Add files via upload
This is basic programme of ATM transaction system 
it's feature is 
Features:
Check Balance: Users can check their current account balance. The balance is initialized to zero and updated based on the transactions made (deposits and withdrawals).

Withdraw Money: Users can withdraw money from their account, with the system checking for sufficient balance. If the balance is insufficient or zero, an appropriate message is displayed to the user, preventing overdraws.

Deposit Money: Users can deposit money into their account. After a successful deposit, the user is asked if they would like to check their updated balance.

Exit: Users can choose to exit the ATM service at any time.

Input Validation: The program includes basic input validation, ensuring that users enter a valid option from the menu (1 to 4). If an invalid input is given, the system prompts the user to enter a correct choice.

Program Flow:
The user is greeted with a welcome message and presented with four options to choose from.
Depending on the user's choice, the corresponding operation is executed (check balance, withdraw money, deposit money, or exit).
After each operation, the program returns to the menu or exits based on the user's input.
Objective:
This project demonstrates fundamental C++ programming concepts such as:

Switch case statements: for handling menu-driven options.
Conditional logic: for ensuring  balance is sufficient or not or etc things .